requirements.txt
flask
Pillow
pycrypto
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
ConfigArgParse==0.13.0
configobj==5.0.6
cryptography==2.2.2
future==0.16.0
idna==2.6
josepy==1.1.0
mock==2.0.0
parsedatetime==2.4
pbr==4.2.0
pycparser==2.18
pymongo==3.7.1
pyOpenSSL==18.0.0
pyparsing==2.2.0
requests==2.18.4
six==1.11.0
urllib3==1.22
